feat: Combine user related validation in one file & Add objectIdValidation file
