+++ routes/user.js
@@ -1,9 +1,10 @@
 const express = require("express");
 
+const { validateToken } = require("../middlewares/userValidation");
 const { getUserMaps } = require("../controllers/user.controller");
 
 const router = express.Router();
 
-router.get("/:id/maps", getUserMaps);
+router.get("/:id/maps", validateToken, getUserMaps);
 
 module.exports = router;
